Move _align_kwcoco_categories to kwcoco_train_detector
