+++ plugins/pytorch/mit_yolo_trainer.py
@@ -17,21 +17,17 @@ import os
 import signal
 import sys
 import time
-import types
 # import math
 
 from .kwcoco_train_detector import KWCocoTrainDetector
 from .kwcoco_train_detector import KWCocoTrainDetectorConfig
+from .kwcoco_train_detector import _align_kwcoco_categories
 from viame.pytorch.utilities import vital_config_update, report_cuda_errors
 
 import scriptconfig as scfg
 import ubelt as ub
 from hydra import compose, initialize_config_dir
 from yolo.lazy import main
-
-
-if types.TYPE_CHECKING:
-    import kwcoco
 
 
 class MITYoloConfig(KWCocoTrainDetectorConfig):
@@ -294,64 +290,6 @@ class MITYoloTrainer( KWCocoTrainDetector ):
             output["train_config.yaml"] = str(train_config_fpath)
 
         return output
-
-
-def _align_kwcoco_categories(
-        dset: kwcoco.CocoDataset,
-        class_list: list[str]
-     ) -> kwcoco.CocoDataset:
-    """
-    Modify (inplace) the CocoDataset so categories are in a specified order,
-    and handle the case of unexpected categories.
-
-    Example:
-        >>> import pytest
-        >>> dset = kwcoco.CocoDataset.demo('vidshapes1')
-        >>> class_list = ['superstar', 'star', 'eff']
-        >>> _align_kwcoco_categories(dset.copy(), class_list)
-        >>> #
-        >>> # Case: dset has category we can't align with, should error
-        >>> with pytest.raises(ValueError):
-        >>>     class_list = ['star', 'eff']
-        >>>     _align_kwcoco_categories(dset.copy(), class_list)
-        >>> #
-        >>> # Case: extra category, should be fine
-        >>> class_list = ['superstar', 'star', 'eff', 'longcat']
-        >>> _align_kwcoco_categories(dset.copy(), class_list)
-    """
-    existing_names = set(dset.categories().lookup('name'))
-    expected_names = set(class_list)
-
-    if len(class_list) != len(expected_names):
-        raise ValueError('class list should not contain duplicates')
-
-    unexpected_names = existing_names - expected_names
-    if unexpected_names:
-        used_cat_ids = set(dset.annots().lookup('category_id', None))
-        used_catnames = {
-            None if cid is None else dset.index.cats[cid].get('name')
-            for cid in used_cat_ids
-        }
-        used_unexpected = used_catnames & unexpected_names
-        if used_unexpected:
-            raise ValueError(
-                f"{dset} contains annotations outside the configured "
-                f"training categories: {sorted(used_unexpected)}"
-            )
-        # Normalize to category ids in case CocoDataset API changes
-        unexpected_cids = [dset.index.name_to_cat[name]['id']
-                           for name in unexpected_names]
-        # Note: keep annots is irrelevant because we error if there are any
-        # annots to remove. However, if we relax that to a warning removing the
-        # annots (or modifying their categories) is the right thing to do, but
-        # that should be handled before it ever gets to this point.
-        dset.remove_categories(unexpected_cids, keep_annots=False)
-
-    for name in class_list:
-        dset.ensure_category(name)
-
-    dset.normalize_category_ids(start_id=1, order=class_list)
-    return dset
 
 
 def __vital_algorithm_register__():
